refactor: 패키지 분리 및 삭제 적용
