+++ src/main/kotlin/racing/domain/Car.kt
@@ -1,4 +1,4 @@
-package racing.car
+package racing.domain
 
 data class Car(
     val name: String = ""
