+++ src/test/kotlin/calculator/CalculatorTest.kt
@@ -1,6 +1,6 @@
 package calculator
 
-import enums.OperatorSign
+import calculator.enums.OperatorSign
 import org.assertj.core.api.Assertions.assertThat
 import org.assertj.core.api.Assertions.assertThatThrownBy
 import org.junit.jupiter.api.Test
